Added the functionallity of login and signup, added the business logic and added the api for it.
In signup, first we are comparing password and confirm password. Then we are checking in the database that user exists or not. Then we are hashing the password using bcrypt and saving the hashed password in database.

In login, we are first checking in the database if user exists or not. Then we are checking the password with the hashed password in the database using bcrypt.compare() and then we are creating token using json web token.